Rename aggregator.py
