+++ lib/pavilion/filters/__init__.py
@@ -1,5 +1,5 @@
 from .filters import (parse_query, get_sort_opts, add_test_filter_args,
     add_series_filter_args, TEST_FILTER_DEFAULTS, SORT_KEYS, SERIES_FILTER_DEFAULTS)
-from .aggregator import AttributeGetter 
+from .attr_getter import AttributeGetter 
 from .validators import validate_int, validate_glob, validate_glob_list, validate_str_list, validate_datetime
 from .errors import FilterParseError
